refactor: 체크인 application 패키지 분리
