+++ src/main/java/com/michelet/restaurantservice/checkin/application/service/RestaurantCheckInCommandService.java
@@ -1,20 +1,20 @@
-package com.michelet.restaurant.application.service.command;
+package com.michelet.restaurantservice.checkin.application.service;
 
 import com.michelet.common.auth.core.enums.UserRole;
 import com.michelet.common.response.ApiResponse;
-import com.michelet.restaurant.application.command.CheckInCommand;
-import com.michelet.restaurant.application.result.CheckInResult;
-import com.michelet.restaurant.domain.exception.CheckInErrorCode;
-import com.michelet.restaurant.domain.exception.CheckInException;
-import com.michelet.restaurant.domain.exception.RestaurantErrorCode;
-import com.michelet.restaurant.domain.exception.RestaurantException;
-import com.michelet.restaurant.domain.model.Restaurant;
-import com.michelet.restaurant.domain.model.RestaurantCheckInLog;
-import com.michelet.restaurant.domain.repository.RestaurantCheckInLogRepository;
-import com.michelet.restaurant.domain.repository.RestaurantRepository;
-import com.michelet.restaurant.infrastructure.client.ReservationClient;
-import com.michelet.restaurant.infrastructure.client.dto.ReservationCheckInRequest;
-import com.michelet.restaurant.infrastructure.client.dto.ReservationCheckInResponse;
+import com.michelet.restaurantservice.checkin.application.command.CheckInCommand;
+import com.michelet.restaurantservice.checkin.application.result.CheckInResult;
+import com.michelet.restaurantservice.checkin.domain.exception.CheckInErrorCode;
+import com.michelet.restaurantservice.checkin.domain.exception.CheckInException;
+import com.michelet.restaurantservice.domain.exception.RestaurantErrorCode;
+import com.michelet.restaurantservice.domain.exception.RestaurantException;
+import com.michelet.restaurantservice.domain.model.Restaurant;
+import com.michelet.restaurantservice.checkin.domain.model.RestaurantCheckInLog;
+import com.michelet.restaurantservice.checkin.domain.repository.RestaurantCheckInLogRepository;
+import com.michelet.restaurantservice.domain.repository.RestaurantRepository;
+import com.michelet.restaurantservice.infrastructure.client.ReservationClient;
+import com.michelet.restaurantservice.infrastructure.client.dto.ReservationCheckInRequest;
+import com.michelet.restaurantservice.infrastructure.client.dto.ReservationCheckInResponse;
 import feign.FeignException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
